+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Drop all Linux capabilities.
+- Expose pod security context configuration.
+- Run as non-root UID by default.
+
 ## [0.0.21] - 2023-12-31
 
 ### Changed
@@ -163,7 +169,6 @@
 <!-- next-url -->
 
 [Unreleased]: https://github.com/agabani/tor-operator/compare/v0.0.21...HEAD
-
 [0.0.21]: https://github.com/agabani/tor-operator/compare/v0.0.20...v0.0.21
 [0.0.20]: https://github.com/agabani/tor-operator/compare/v0.0.19...v0.0.20
 [0.0.19]: https://github.com/agabani/tor-operator/compare/v0.0.18...v0.0.19